Move logger into separate .h/.cpp files
