+++ plugins/soundsourcemediafoundation/soundsourcemediafoundation.cpp
@@ -5,7 +5,7 @@
 #include <propvarutil.h>
 
 #include "util/sample.h"
-#include "util/logging.h"
+#include "util/logger.h"
 
 namespace {
