+++ plugins/soundsourcewv/SConscript
@@ -19,7 +19,7 @@ wv_sources = [
     "util/samplebuffer.cpp",
     "util/singularsamplebuffer.cpp",
     "util/sample.cpp",
-    "util/logging.cpp",
+    "util/logger.cpp",
     "track/trackmetadata.cpp",
     "track/trackmetadatataglib.cpp",
     "track/tracknumbers.cpp",
