+++ plugins/soundsourcewv/soundsourcewv.cpp
@@ -2,7 +2,7 @@
 
 #include "soundsourcewv.h"
 
-#include "util/logging.h"
+#include "util/logger.h"
 
 namespace mixxx {
