+++ src/sources/audiosource.cpp
@@ -1,7 +1,7 @@
 #include "sources/audiosource.h"
 
 #include "util/sample.h"
-#include "util/logging.h"
+#include "util/logger.h"
 
 namespace mixxx {
