+++ src/sources/soundsourceopus.cpp
@@ -1,6 +1,6 @@
 #include "sources/soundsourceopus.h"
 
-#include "util/logging.h"
+#include "util/logger.h"
 
 namespace mixxx {
