+++ src/sources/soundsourcesndfile.cpp
@@ -2,7 +2,7 @@
 
 #include "sources/soundsourcesndfile.h"
 
-#include "util/logging.h"
+#include "util/logger.h"
 
 namespace mixxx {
