+++ src/util/logging.h
@@ -1,7 +1,7 @@
 #ifndef MIXXX_UTIL_LOGGING_H
 #define MIXXX_UTIL_LOGGING_H
 
-#include <QtDebug>
+#include <QDir>
 
 
 namespace mixxx {
@@ -17,7 +17,7 @@ enum class LogLevel {
 class Logging {
   public:
     // These are not thread safe. Only call them on Mixxx startup and shutdown.
-    static void initialize(const QString& settingsPath,
+    static void initialize(const QDir& settingsDir,
                            LogLevel logLevel,
                            bool debugAssertBreak);
     static void shutdown();
@@ -40,49 +40,6 @@ class Logging {
     Logging() = delete;
 
     static LogLevel s_logLevel;
-};
-
-class Logger final {
-public:
-    Logger() = default;
-    explicit Logger(const char* logContext);
-    explicit Logger(const QString& logContext);
-
-    QDebug log(QDebug stream) const {
-        return stream << m_preambleChars.data();
-    }
-
-    QDebug debug() const {
-        return log(qDebug());
-    }
-
-    bool debugEnabled() const {
-        return Logging::debugEnabled();
-    }
-
-    QDebug info() const {
-#if QT_VERSION >= QT_VERSION_CHECK(5, 0, 0)
-        return log(qInfo());
-#else
-        // Qt4 does not support log level Info, use Debug instead
-        return debug();
-#endif
-    }
-
-    bool infoEnabled() const {
-        return Logging::infoEnabled();
-    }
-
-    QDebug warning() const {
-        return log(qWarning());
-    }
-
-    QDebug critical() const {
-        return log(qCritical());
-    }
-
-private:
-    QByteArray m_preambleChars;
 };
 
 }  // namespace mixxx
